Change height_of_device type from int to float

+++ src/ralph_assets/models_assets.py
@@ -174,7 +174,7 @@ class AssetModel(
         blank=True,
         default=0,
     )
-    height_of_device = models.IntegerField(
+    height_of_device = models.FloatField(
         verbose_name="Height of device",
         blank=True,
         default=0,
